Seperated homepage and main app widget classes
